// deploy/restpe-contabilidad-front-end/src/app/modules/rrhh/domain/models/distribucion-costos.entity.ts
export interface DistribucionCostosEntity {
  distribucion_costos_periodo: string;
  distribucion_costos_trabajador: string;
  distribucion_costos_sucursal: string;
  distribucion_costos_centro_costo: string;
  distribucion_costos_canal: string;
  distribucion_costos_tipo_distribucion: string;
  distribucion_costos_aplicado: number;
  distribucion_costos_monto_asig: number;
}